rename all the things to debezium as suggested by Gunnar
