+++ kstreams/poc-ddd-aggregates/pom.xml
@@ -3,7 +3,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.github.hpgrahsl.kafka</groupId>
+    <groupId>io.debezium.examples</groupId>
     <artifactId>poc-ddd-aggregates</artifactId>
     <version>0.1-SNAPSHOT</version>
