--- kstreams/poc-ddd-aggregates/src/main/java/com/github/hpgrahsl/kafka/model/EventType.java
@@ -1,3 +1,0 @@
-package com.github.hpgrahsl.kafka.model;
-
-public enum EventType { UPSERT, DELETE }
