+++ kstreams/poc-ddd-aggregates/src/main/java/io/debezium/examples/aggregation/model/EventType.java
@@ -1,0 +1,3 @@
+package io.debezium.examples.aggregation.model;
+
+public enum EventType { UPSERT, DELETE }
